Remove IAppxFile from StreamBase. Most of the derived classed of StreamBase don't care about if it is an IAppxFile or not and we end up with a lot of them just returning E_NOTIMPL. This is also confusing sometimes a stream can QI for IAppxFile and succeed, but if we ever get an outside stream and asume that is an appx file we will have problems.

IAppxFileInternal had nothing to do with AppxFiles, so I remove it. Its only functionality was to return the size of the stream in the zip container to validate the information of the block map. IStreamInternal::GetSizeOnZip provides this functionality now.

Because the vtables are smaller, this change also reduce the size of the binary ~7k

Related work items: #18752968, #18853644

+++ src/msix/PAL/Signature/Win32/SignatureValidator.cpp
@@ -452,13 +452,13 @@ namespace MSIX
         ThrowHrIfFailed(stream->Seek(li, StreamBase::Reference::START, &uli));
 
         ULONG actualRead = 0;
-        ThrowHrIfFailed(stream->Read(p7x.data(), p7x.size(), &actualRead));
+        ThrowHrIfFailed(stream->Read(p7x.data(), static_cast<ULONG>(p7x.size()), &actualRead));
         ThrowErrorIf(Error::SignatureInvalid,
             ((actualRead != p7x.size() || (*(DWORD*)p7x.data() != P7X_FILE_ID))),
             "Failed to read p7x or p7x header mismatch");
 
         BYTE *p7s = p7x.data() + sizeof(DWORD);
-        std::uint32_t p7sSize = p7x.size() - sizeof(DWORD);
+        std::uint32_t p7sSize = static_cast<uint32_t>(p7x.size() - sizeof(DWORD));
 
         // Decode the ASN.1 structure
         CRYPT_CONTENT_INFO* contentInfo = nullptr;
